Fix for when Accept header mime-types include spaces on sides.
This problem is mainly caused by some crazy accept header that's
sent 50% of the time in IE, but it may be caused by other
browsers that add additional whitespace.  Not sure if the standard
is clear on whether extra whitespace can or can't exist, but
either way this solves it for non-compliant browsers.